chore: addressed PR comments
